DROOLS-3831: Implementing review commentaries
